requirements.txt
cffi==1.9.1
cryptography==1.7.2
dj-database-url==0.4.2
Django==1.10.4
django-crispy-forms==1.6.1
django-tagging==0.4.5
enum34==1.1.6
gunicorn==19.7.1
idna==2.4
ipaddress==1.0.18
mysqlclient==1.3.9
olefile==0.44
Pillow==4.0.0
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.17
pyOpenSSL==0.15.1
python-memcached==1.58
requests==2.13.0
six==1.10.0
stripe==1.46.0
whitenoise==3.3.0
